Keep track of "session" and reset state when "session" has changed
Just save a UUIDv4 to the command station as a session.
Controlling an engine also adds the session to the state.
If the session of an engine and the current session don't match, we don't know about the physical state of the engine and we reset the state (only clientside for now)